📁 (#62) - add content folder & move content to content folder
