+++ nutella/src/components/Result/Content/SubmitResult/index.tsx
