+++ nutella/src/components/Result/Content/SubmitResult/styles.ts
